Unify Namespace parsing code
