+++ syntax/namespace.rs
@@ -1,5 +1,13 @@
+use crate::syntax::ident;
+use quote::IdentFragment;
 use std::fmt::{self, Display};
 use std::slice::Iter;
+use syn::parse::{Parse, ParseStream, Result};
+use syn::{Path, Token};
+
+mod kw {
+    syn::custom_keyword!(namespace);
+}
 
 #[derive(Clone)]
 pub struct Namespace {
@@ -7,12 +15,31 @@ pub struct Namespace {
 }
 
 impl Namespace {
-    pub fn new(segments: Vec<String>) -> Self {
-        Namespace { segments }
+    pub fn none() -> Self {
+        Namespace {
+            segments: Vec::new(),
+        }
     }
 
     pub fn iter(&self) -> Iter<String> {
         self.segments.iter()
+    }
+}
+
+impl Parse for Namespace {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let mut segments = Vec::new();
+        if !input.is_empty() {
+            input.parse::<kw::namespace>()?;
+            input.parse::<Token![=]>()?;
+            let path = input.call(Path::parse_mod_style)?;
+            for segment in path.segments {
+                ident::check(&segment.ident)?;
+                segments.push(segment.ident.to_string());
+            }
+            input.parse::<Option<Token![,]>>()?;
+        }
+        Ok(Namespace { segments })
     }
 }
 
@@ -23,6 +50,12 @@ impl Display for Namespace {
             f.write_str("$")?;
         }
         Ok(())
+    }
+}
+
+impl IdentFragment for Namespace {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        Display::fmt(self, f)
     }
 }
